Refactor configuration file settings: move reveal/edit actions to General settings tab and add runtime state reset options
