+++ .changeset/20260613005436-added-reset-runtime-state-and-restart-clearing-s.md
@@ -5,4 +5,4 @@
 
 Added Reset Runtime State and Restart Clearing State developer menu items to the status bar menu (visible when Developer Mode is enabled). Both actions require confirmation when triggered from the status bar menu or command palette. The restart dialog includes an Enable Tracing checkbox that starts runtime trace capture as early as possible during app bootstrap.
 
-Moved Reveal Config Folder and Edit settings.toml from the status bar menu to a dedicated Config Files section in Settings.
+Moved Reveal Config Folder and Edit settings.toml from the status bar menu into Settings → General.
