+++ Sources/Nehir/UI/SettingsView.swift
@@ -37,6 +37,10 @@ struct GeneralSettingsTab: View {
     @Bindable var controller: WMController
     var cliManager: AppCLIManager?
 
+    @State private var resultStatus: SettingsFileStatus?
+    @State private var resultVisible = false
+    @State private var dismissTask: DispatchWorkItem?
+
     var body: some View {
         Form {
             Section("Appearance") {
@@ -93,8 +97,52 @@ struct GeneralSettingsTab: View {
                 CLISettingsSection(cliManager: cliManager)
             }
 
+            ConfigurationFilesSection(settings: settings) { action in
+                performAction(action)
+            }
+
         }
         .formStyle(.grouped)
+        .overlay {
+            if resultVisible, let status = resultStatus {
+                VStack {
+                    Spacer()
+                    HStack(spacing: 6) {
+                        Image(systemName: status.icon)
+                            .foregroundStyle(status.color)
+                        Text(status.message)
+                            .font(.caption)
+                            .foregroundStyle(.secondary)
+                    }
+                    .padding(.horizontal, 12)
+                    .padding(.vertical, 6)
+                    .background(.regularMaterial, in: Capsule())
+                    .transition(.move(edge: .bottom).combined(with: .opacity))
+                }
+                .padding(.bottom, 16)
+            }
+        }
+        .onDisappear {
+            dismissTask?.cancel()
+            dismissTask = nil
+        }
+    }
+
+    private func performAction(_ action: SettingsFileAction) {
+        dismissTask?.cancel()
+        withAnimation { resultVisible = false }
+        do {
+            let status = try SettingsFileWorkflow.perform(action, settings: settings)
+            resultStatus = status
+        } catch {
+            resultStatus = .error(error.localizedDescription)
+        }
+        withAnimation { resultVisible = true }
+        let task = DispatchWorkItem {
+            withAnimation { self.resultVisible = false }
+        }
+        dismissTask = task
+        DispatchQueue.main.asyncAfter(deadline: .now() + 2, execute: task)
     }
 }
